+++ frontend/src/pages/HumanChatPage/ChattingPage.tsx
@@ -5,6 +5,7 @@ import { IconButton } from '@components/Buttons';
 import CamContainer from '@components/CamContainer';
 
 import { useControllMedia } from '@business/hooks/useWebRTC/useControllMedia';
+import { useStreamVideoRef } from '@business/hooks/useWebRTC/useStreamVideoRef';
 import WebRTC from '@business/services/WebRTC';
 
 import type { OutletContext } from './HumanChatPage';
@@ -12,8 +13,8 @@ import { useChattingPageCreateJoinRoomPasswordPopup } from './useChattingPageCre
 
 export default function ChattingPage() {
   const {
-    localVideoRef,
-    remoteVideoRef,
+    // localVideoRef,
+    // remoteVideoRef,
     tarotButtonDisabled,
     tarotButtonClick,
     enableSideBar,
@@ -21,27 +22,15 @@ export default function ChattingPage() {
     unblockGoBack,
   }: OutletContext = useOutletContext();
 
+  useChattingPageCreateJoinRoomPasswordPopup({ unblockGoBack, enableSideBar });
+  const { localVideoRef, remoteVideoRef } = useStreamVideoRef();
   const { toggleAudio, toggleVideo, changeMyVideoTrack } = useControllMedia({ localVideoRef });
-  const webRTC = WebRTC.getInstace();
 
   useEffect(() => {
     if (joined) {
       changeMyVideoTrack();
     }
   }, [joined]);
-
-  useEffect(() => {
-    const existRemoteVideo = remoteVideoRef.current;
-    const existRemoteStream = webRTC.remoteStream;
-    const remoteStreamChanged = (remoteVideoRef.current?.srcObject as MediaStream)?.id !== webRTC.remoteStream?.id;
-    if (!existRemoteVideo || !existRemoteStream || !remoteStreamChanged) {
-      return;
-    }
-
-    remoteVideoRef.current.srcObject = webRTC.remoteStream as MediaStream;
-  }, [remoteVideoRef.current]);
-
-  useChattingPageCreateJoinRoomPasswordPopup({ unblockGoBack, enableSideBar });
 
   const navigate = useNavigate();
   const goSettingPage = () => navigate('setting');
